LocalInternet/services/bank/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
jinja2
python-multipart
requests
python-jose[cryptography]
